+++ src/parser/check_func1.c
@@ -6,14 +6,14 @@
 /*   By: jaelee <jaelee@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/23 17:13:50 by jaelee            #+#    #+#             */
-/*   Updated: 2019/01/30 14:12:55 by jaelee           ###   ########.fr       */
+/*   Updated: 2019/02/03 22:26:09 by aamadori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parser.h"
 #include "get_next_line.h"
 
-size_t  is_nbr_ants(char *line, t_lemin *info)
+int  is_nbr_ants(char *line, t_lemin *info)
 {
 	(void)info;
 	if (ft_str_is_digit(line) && ft_atoi(line) > 0)
@@ -21,7 +21,7 @@ size_t  is_nbr_ants(char *line, t_lemin *info)
 	return (0);
 }
 
-size_t  is_start(char *line, t_lemin *info)
+int  is_start(char *line, t_lemin *info)
 {
 		(void)info;
 		if(!ft_strcmp("##start", line))
@@ -29,7 +29,7 @@ size_t  is_start(char *line, t_lemin *info)
 		return (0);
 }
 
-size_t  is_end(char *line, t_lemin *info)
+int  is_end(char *line, t_lemin *info)
 {
 		(void)info;
 		if(!ft_strcmp("##end", line))
@@ -37,7 +37,7 @@ size_t  is_end(char *line, t_lemin *info)
 		return (0);
 }
 
-size_t  is_command(char *line, t_lemin *info)
+int  is_command(char *line, t_lemin *info)
 {
 		(void)info;
 		if(line[0] == '#' && line[1] == '!')
